Add basic styling
The recipes looked a bit dull with just raw HTML and no CSS. Now that
I've been introduced to the joys of styling websites, I can add some
basic CSS styles to this project. The aim is not really to make it
look good, but rather to practice the CSS lessons.